chore: move RedisIdentifierBuilder to Infrastructure
